Make vote casting ack a modal dialog
Overwhelming feedback from VPSI test users said that they
wished that the "your vote has been cast" message was more
prominent.

Students user testers, for some reason, never commented
on this during user testing.

Perhaps it is a generational difference.

Anyway, the change is easy to make, so let's bow to the
wisdom of the elders.