Sleep before doing the lookup to workaround IE11 problem
IE11 has a limit on the max number of websockets that can be active,
and so on ballots with 7 people, the last one is always not looked
up. The hacky solution is to put a sleep on the front of the lookups
and when i > 5 (i.e. the 7th, 8th, 9th candidate is being looked up)
then bump the sleep time from 0 to 2 seconds. This postpones the 7th
lookup until the first one has finished, keeping us under the limit
of concurrent lookups.