Update to Vue 3 properly
Remove the whacky build/config stuff, and instead use
Vue's defaults as much as possible.

Simplify things as much as possible, i.e. use Material Icons
from Google, remove CSS that wasn't working anymore, etc.

Change the protocol to talk to the worker to work again, because
Vue 3 did not like how we were talking to it before.

Switch to yarn because that's what Vue 3 defaults to, and it's faster.