Testing improvments
Fix reconstruction to only talk to the leader. As long as
the public.toml file only has the leader in it, this was never
really a problem, but better like this in case the public.toml
has the whole roster, and the roster is modified on the master
chain.

Also: A useful feature for testing, ability to inpersonate
another user. (Would never be turned on in prod.)